Move commonly used code to base strategy
In all the strategies we've written so far, the following hasn't
really changed:
- adjustPosition
- liquidatePosition
- the second half of harvestCore

In this commit we pull all this common code into the base strategy
while still allowing it be overridden (all marked "virtual"). For
consumers of the base strategy, it is only necessary to implement
_deposit(), _withdraw(), and a simplified _harvestCore() for
processing rewards.

Note that liquidateAllPositions is still unimplemented in the base
because it could be confusing to expect a consumer of base strategy
to know how to handle withdraw(maxInt). Or they may need to do
additional processing besides simply calling withdraw().